Move buildRequest to separate api package
In order to simplify the builder package (to fix import cycle issue),
move the BuildRequest struct into a common API package that can
be imported by the cmd, builder, and downloader. This significantly
simplifies the gomocks for the builder package, and allows the
Downloader to return to the builder package.