Add generate-manifests command
In an effort to make stepping through and debugging easier, add a new
`generate-manifests` command. This command takes a namespace and
optionally auth and an appregistry endpoint, parses the binaries and
pushes them into a local directory.

Additionally, refactored the build handler into the apis package in an
effort to be able to reuse the rest of the build functions without it.

Also refactored the downloader into a separate package in an effort to
reduce the import footprint of the generate handler.